# back-end/tel_openapi.yaml
openapi: "3.0.2"
info:
  description: Control Center API description
  title: Control Center
  version: 1.0.0
paths:
...
  /admin/users:
    get:
      tags:
        - AdminUsers
      summary: Get all users
      operationId: get_admin_users
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userArray"
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint

    post:
      tags:
        - AdminUsers
      summary: Add a new user
      operationId: post_admin_users
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/user"
        required: true
      responses:
        '200':
          description: Successful addition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        '500':
          description: Internal server error
        '400':
          description: Bad request
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
  /admin/users/{userId}:
    get:
      tags:
        - AdminUsersId
      summary: Get a specific user
      operationId: get_admin_users_byid
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          schema:
            $ref: "#/components/schemas/userId"
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
          
    patch:
      tags:
        - AdminUsersId
      summary: Update a user
      operationId: patch_admin_users_byid
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          schema:
            $ref: "#/components/schemas/userId"
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/user"
      responses:
        '200':
          description: Successful update
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
          
    delete:
      tags:
        - AdminUsersId
      summary: Delete a user
      operationId: delete_admin_users_byid
      parameters:
        - name: userId
          in: path
          description: ID of user
          required: true
          schema:
            $ref: "#/components/schemas/userId"
      responses:
        '200':
          description: Successful deletion
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
  /agencies:
    get:
      tags:
        - agencies
      summary: Get all agencies available
      operationId: get_agencies
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/agencyArray"
        '404':
          description: Not found
        '401':
          description: Not Authorized for this operation
        '500':
          description: Internal server error
    
    post:
      tags:
        - agencies
      summary: Submit a new agency
      operationId: post_agencies
      requestBody:
        content:
          application/json:   
            schema:
              $ref: "#/components/schemas/agency"
        required: true
      responses:
        '200':
          description: Successful submition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/agency"
        '500':
          description: Internal server error
        '401':
          description: Not Authorized for this operation
        '400':
          description: Bad request
        '403':
          description: Forbidden to use this endpoint
  /agencies/{agencyId}:
    get:
      tags:
        - agenciesId
      summary: Get an agency
      operationId: get_agencies_byid
      parameters:
        - name: agencyId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/agencyId"
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/agency"
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
    delete:
      tags:
        - agenciesId
      summary: Delete an agency
      operationId: delete_agencies_byid
      parameters:
        - name: agencyId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/agencyId"
      responses:
        '200':
          description: Successful deletion
        '404':
          description: Not found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
    
    put:
      tags:
        - agenciesId
      summary: Update an agency
      operationId: put_agencies_byid
      parameters:
        - name: agencyId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/agencyId"
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/agency"
      responses:
        '200':
          description: Successful update
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
...
components:
  schemas:
    agencyId:
      type: string
      description: String containing the agency id
    userId:
      type: string
      description: String containing the user id
...
    user:
      type: object
      required:
        - userId
      properties:
        userId:
          $ref: "#/components/schemas/userId"
        username:
          type: string
          description: Defines unique username, admin has always 'admin' as username
        email:
          type: string
          description: Defines unique email
        password:
          type: string
          minimum: 8
          maximum: 31
          pattern: '^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$'
        role:
          type: string
          description: Role in agency
        agencyId:
          type: string
          description: ID to identify the agency of the user if applicable.
        firstName:
          type: string
        lastName:
          type: string
        age:
          type: number
        gender:
          type: string
        isHeadOfAgency:
          type: boolean
          description: Defines users that represent their corresponding agencies (equals false for normal users)
        x:
          type: number
          description: longtitude
        y:
          type: number
          description: latitude
        isAvailable:
          type: boolean
          description: Used only by normal users (not admin,coordinator,head of agency)
    userArray:
      type: array
      items:
        $ref: "#/components/schemas/user"
...
    agency:
      type: object
      required:
        - agencyId
      properties:
        agencyId:
          $ref: "#/components/schemas/agencyId"
        name:
          type: string
          description: Unique name of agency
        userId:
          $ref: "#/components/schemas/userId"
        address:
          type: string
          description: Physical address of agency (not coordinates)
        users:
          $ref: "#/components/schemas/userArray"
    agencyArray:
      type: array
      items:
        $ref: "#/components/schemas/agency"
...